An Improved N-gram Distance for Names Matching
N-gram distance (N-DIST) was basically developed by Kondrak’s lately to measure the distance between two strings. It was found that this distance can be computed by an smart dynamic programming procedure. The N-DIST has played important roles in a wide array of applications due to its representational efficacy and computational efficiency. To effect a more sensible, distance measure, the normalized edit distance was proposed. Many algorithms and studies have been dedicated along this line with impressive performances in last years. There is, however, a fundamental problem with the original definition of N-DIST that has remained without improved: its context-free nature. In determining the possible actions, i.e., deletion, insertion, transposition and substitution, consider work was given to the local behaviors of the string question that indeed encompass great amount of useful information concerning its content. This proposed framework, two operations are developed: The original N-DIST algorithm does not consider the transposition operations, and the algorithm has fixed the cost of insertion and deletion operations. In addition, the proposed E-N-DIST algorithm computes the costs of substitution and transposition operations is dependent on 2n+1 -1 states while the original N-DIST algorithm has been only dependent on 2n states. In this paper, the experiments carried out show the E-N-DIST algorithm which gives a sort of results which are more accurate than the algorithms under discussion.